Fix warning: CA1901  P/Invoke declarations should be portable    As it is declared in your code, the return type of P/Invoke 'RichTextBoxXhtmlSupportExtension.NativeMethods.SendMessage(HandleRef, int, IntPtr, IntPtr)' will be 4 bytes wide on 64-bit platforms. This is not correct, as the actual native declaration of this API indicates it should be 8 bytes wide on 64-bit platforms. Consult the MSDN Platform SDK documentation for help determining what data type should be used instead of 'int'.  GitUI   RichTextBoxXhtmlSupportExtension.cs 333

+++ GitUI/Editor/RichTextBoxXhtmlSupportExtension.cs
@@ -33,7 +33,7 @@ namespace GitUI.Editor.RichTextBoxExtension
         private static int BeginUpdate(HandleRef handleRef)
         {
             // Prevent the control from raising any events.
-            int oldEventMask = NativeMethods.SendMessage(handleRef,
+            int oldEventMask = (int) NativeMethods.SendMessage(handleRef,
                 NativeMethods.EM_SETEVENTMASK, IntPtr.Zero, IntPtr.Zero);
 
             // Prevent the control from redrawing itself.
@@ -330,7 +330,7 @@ namespace GitUI.Editor.RichTextBoxExtension
             internal const int LF_FACESIZE = 32;
 
             [DllImport("user32", CharSet = CharSet.Auto)]
-            internal static extern int SendMessage(HandleRef hWnd,
+            internal static extern IntPtr SendMessage(HandleRef hWnd,
                 int msg,
                 IntPtr wParam,
                 IntPtr lParam);
